backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
PyPDF2==3.0.1
google-generativeai==0.3.2
python-multipart==0.0.6
pydantic==2.5.3
aiofiles==23.2.1